+++ tests/python_tests/test_sampling.py
@@ -3,14 +3,13 @@
 
 import sys
 import pytest
-import sys
 from dataclasses import dataclass
 from pathlib import Path
-from openvino_genai import GenerationConfig, StopCriteria
 from typing import List, TypedDict
 
-from common import run_llm_pipeline_with_ref, run_llm_pipeline
+from openvino_genai import GenerationConfig, StopCriteria
 
+from common import run_llm_pipeline_with_ref, run_llm_pipeline
 from utils.hugging_face import get_hugging_face_models, convert_models
 
 @pytest.mark.precommit
